backend/requirements.minimal.txt
# BL Genius - Backend Dependencies MINIMAL (Build rapide)

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6

# Database (binary = pas besoin de gcc)
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Celery
celery==5.3.6

# YOLO and tracking
ultralytics==8.1.0
supervision==0.19.0

# Computer vision - HEADLESS (pas besoin de libGL)
opencv-python-headless==4.8.0.74
numpy==1.26.3

# Machine Learning (K-Means)
scikit-learn==1.4.0

# Interpolation
scipy==1.12.0

# Training
roboflow==1.1.23

# Utilities
pillow==10.2.0
tqdm==4.66.1

# Authentification JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation des fichiers (MIME type)
python-magic==0.4.27

# Rate Limiting
slowapi==0.1.9

# Validation Email
email-validator==2.1.0

# Hashing sécurisé
bcrypt==4.1.2

# Pour les tokens sécurisés
cryptography==41.0.7

# YouTube download
yt-dlp==2023.12.30

# WebSocket
websockets==12.0
